Use maven-process-plugin to start agent in integration tests
To simplify integration tests in preparation for #64
