+++ jmxtrans2-agent/src/test/resources/project-info.properties
@@ -21,7 +21,4 @@
 # THE SOFTWARE.
 #
 
-project.version=${project.version}
-project.build.finalName=${project.build.finalName}
 project.build.directory=${project.build.directory}
-project.basedir=${project.basedir}
